[PERFECTIVE] Fixed rest of the compiler warnings on g++.
git-svn-id: svn://svn.code.sf.net/p/kactus2/code/trunk@1281 f0aa6db5-7f46-43cc-a8a5-a31efdaafa67

+++ models/ApiFunction.cpp
@@ -304,7 +304,7 @@ int ApiFunction::getParamCount() const
 //-----------------------------------------------------------------------------
 // Function: ApiFunction::generateToolTipText()
 //-----------------------------------------------------------------------------
-void ApiFunction::generateToolTipText(unsigned int paramIndex, QString& text) const
+void ApiFunction::generateToolTipText(int paramIndex, QString& text) const
 {
     text.clear();
 
@@ -335,7 +335,7 @@ void ApiFunction::generateToolTipText(unsigned int paramIndex, QString& text) co
 
     text += ")<br>" + desc_;
     
-    if (paramIndex < static_cast<unsigned int>(params_.size()))
+    if (paramIndex < params_.size())
     {
         text += "<br><b>Parameter:</b> " + params_.at(paramIndex)->getDescription();
     }
